\name{banco}
\alias{banco}
\title{
Banco de dados
}
\description{
Banco de dados com varias variaveis
}
\usage{
data(banco)
}

\arguments{
  \item{banco}{nome do banco de dados}
}
\details{Este banco de dados contem varias variaveis do tipo qualitativa e quantitativa.
}
\value{Sao variaveis deste banco de dados:
\item{idMM}{identificacao do aluno que forneceu os dados para o banco de dados}
\item{turmaMM}{turma em que o aluno foi alocado (A=1 ou B=2)}
\item{sexoMM}{1 se feminino e 2 se masculino}
\item{idadeMM}{idade em anos completos}
\item{alturaMM}{altura em metros}
\item{pesoMM}{peso em quilogramas}
\item{filhosMM}{numero de filhos na familia}
\item{fumaMM}{tem habito de fumar, sim=0 ou nao=1}
\item{tolerMM}{tolerancia ao cigarro: (I=1) indiferente, (P=2) incomoda pouco e (M=3) incomoda muito}
\item{exerMM}{quantas horas completas de atividade fisica, por semana}
\item{cineMM}{numero de vezes que vai ao cinema por semana}
\item{opcineMM}{opiniao a respeito das salas de cinema Da cidade: (B=1) regular a boa e (M=2) muito boa}
\item{tvMM}{horas gastas assistindo TV, por semana}
\item{optvMM}{opiniao a respeito da qualidade da programacao na TV: (R=1) ruim, (M=2) media, (B=3) boa e (N=10) nao sabe}

}
\references{
MAGALHAES, Marcos Nascimento; LIMA, Antonio Carlos Pedroso de. Nocoes de probabilidade e estatistica. 7 ed. Sao Paulo: EdUSP, 2010. 408p.

Banco de dados obtido em: \url{https://www.ime.usp.br/~noproest/doku.php?id=dados}

}
\author{Adriana Dias,
\email{adriana.dias@unifal-mg.edu.br}

Denismar Alves Nogueira,
\email{denismar.nogueira@unifal-mg.edu.br}

Flavio Bittencourt,
\email{flavio.bittencourt@unifal-mg.edu.br}

Natalia da Silva Martins Fonseca,
\email{natalia.martins@unifal-mg.edu.br}
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{



}
