\name{banco4}
\alias{banco4}
\title{Banco de dados
}
\description{
Banco de dados com duas variaveis
}
\usage{
data(banco4)
}

\arguments{
  \item{banco4}{nome do banco de dados}

}
\details{Este banco de dados contem quatro variaveis.
}
\value{Sao variaveis deste banco de dados:
\item{ano}{Ano do calendario em que os dados foram coletados}
\item{X}{Ano convertido para uma escala que se inicia em 1}
\item{Y}{Taxa de fatalidade por 1000 empregados}
\item{lnX}{Logaritmo natural da variavel X}
}
\references{
PAGANO, M.; GAUVREAU, K. Principios de bioestatistica. 2. ed. Sao Paulo: Cengage Learning, 2012. p. 506.
}
\author{
Adriana Dias,
\email{adriana.dias@unifal-mg.edu.br}

Denismar Alves Nogueira,
\email{denismar.nogueira@unifal-mg.edu.br}

Flavio Bittencourt,
\email{flavio.bittencourt@unifal-mg.edu.br}

Natalia da Silva Martins Fonseca,
\email{natalia.martins@unifal-mg.edu.br}
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
