\name{figuras.pict}
\alias{figuras.pict}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Lista de figuras para o pictograma
}
\description{
Exibe a lista de figuras para uso no pictograma
}
\usage{
figuras.pict(figs)
}

\arguments{
  \item{figs}{exibe os nomes das figuras disponiveis no pacote para uso no pictograma.}
}
\details{Permite visualizar os nomes das figuras disponiveis para uso no pictograma.

}
\value{Retorna visualizar os nomes das figuras disponiveis para uso no pictograma.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{graf.pic}}
}
\examples{
library(classRoom)

figuras.pict(figs)
}
