\name{th.fisher}
\alias{th.fisher}

\title{
Teste exato de Fisher
}
\description{
Apresenta o valor-p do teste exato de Fisher
}
\usage{
th.fisher(r,c,tipo="D")
}

\arguments{
\item{r}{valor que corresponde ao numero de linhas da tabela}
\item{c}{valor que corresponde ao numero de colunas da tabela}
\item{tipo}{tipo de hipotese em teste: "B" - teste bilateral, mas ha outras opcoes: "E" para o teste unilateral a esquerda e "D" (\emph{default}) para o teste unilateral a direita.}
 }

\details{
Permite obter o valor-p do teste exato de Fisher
}
\value{
Retorna o valor-p do teste exato de Fisher
}
\references{

%atualizar a bibliografia

OLIVEIRA et al. Introducao a Estatistica. 2. ed. Lavras: Editora UFLA, 2014. p. 461.

PAGANO, M.; GAUVREAU, K. Principios de bioestatistica. 2. ed. Sao Paulo: Cengage Learning, 2012. p. 506.
}
\author{
Adriana Dias,
\email{adriana.dias@unifal-mg.edu.br}

Denismar Alves Nogueira,
\email{denismar.nogueira@unifal-mg.edu.br}

Flavio Bittencourt,
\email{flavio.bittencourt@unifal-mg.edu.br}

Natalia da Silva Martins Fonseca,
\email{natalia.martins@unifal-mg.edu.br}
}
\note{
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{



}
